+++ src/main/java/roomescape/api/reservation/ReservationController.java
@@ -12,7 +12,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.domain.reservation.CreateReservation;
 import roomescape.domain.reservation.Reservation;
-import roomescape.domain.reservation.ReservationRepository;
+import roomescape.infra.reservation.ReservationRepository;
 
 @RestController
 @RequestMapping("reservations")
