+++ src/main/java/roomescape/api/theme/ThemeController.java
@@ -12,7 +12,7 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.domain.theme.CreateTheme;
 import roomescape.domain.theme.Theme;
-import roomescape.domain.theme.ThemeRepository;
+import roomescape.infra.theme.ThemeRepository;
 
 @RestController
 @RequestMapping("themes")
