+++ src/main/java/roomescape/api/time/ReservationTimeController.java
@@ -3,7 +3,6 @@ package roomescape.api.time;
 import java.util.List;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.Mapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -11,7 +10,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.domain.time.CreateReservationTime;
 import roomescape.domain.time.ReservationTime;
-import roomescape.domain.time.ReservationTimeRepository;
+import roomescape.infra.time.ReservationTimeRepository;
 
 @RestController
 @RequestMapping("times")
