+++ src/main/java/roomescape/infra/theme/ThemeRowMapper.java
@@ -1,9 +1,10 @@
-package roomescape.domain.theme;
+package roomescape.infra.theme;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Component;
+import roomescape.domain.theme.Theme;
 
 @Component
 class ThemeRowMapper implements RowMapper<Theme> {
