+++ src/main/java/roomescape/infra/time/ReservationTimeRepository.java
@@ -1,4 +1,4 @@
-package roomescape.domain.time;
+package roomescape.infra.time;
 
 import java.sql.PreparedStatement;
 import java.sql.Time;
@@ -6,6 +6,8 @@ import java.util.List;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.stereotype.Repository;
+import roomescape.domain.time.CreateReservationTime;
+import roomescape.domain.time.ReservationTime;
 
 @Repository
 public class ReservationTimeRepository {
